+++ context.go
@@ -288,7 +288,7 @@ const (
 // Raw strings are with a terminating NUL character. For convenience, use ContextInfoString().
 //
 // See also: https://registry.khronos.org/OpenCL/sdk/3.0/docs/man/html/clGetContextInfo.html
-func ContextInfo(context Context, paramName ContextInfoName, paramSize uint, paramValue unsafe.Pointer) (uint, error) {
+func ContextInfo(context Context, paramName ContextInfoName, paramSize uintptr, paramValue unsafe.Pointer) (uintptr, error) {
 	sizeReturn := C.size_t(0)
 	status := C.clGetContextInfo(
 		context.handle(),
@@ -299,7 +299,7 @@ func ContextInfo(context Context, paramName ContextInfoName, paramSize uint, par
 	if status != C.CL_SUCCESS {
 		return 0, StatusError(status)
 	}
-	return uint(sizeReturn), nil
+	return uintptr(sizeReturn), nil
 }
 
 // ContextInfoString is a convenience method for ContextInfo() to query information values that are string-based.
@@ -307,7 +307,7 @@ func ContextInfo(context Context, paramName ContextInfoName, paramSize uint, par
 // This function does not verify the queried information is indeed of type string. It assumes the information is
 // a NUL terminated raw string and will extract the bytes as characters before that.
 func ContextInfoString(context Context, paramName ContextInfoName) (string, error) {
-	return queryString(func(paramSize uint, paramValue unsafe.Pointer) (uint, error) {
+	return queryString(func(paramSize uintptr, paramValue unsafe.Pointer) (uintptr, error) {
 		return ContextInfo(context, paramName, paramSize, paramValue)
 	})
 }
